+++ src/modules/bills/entities/income.entity.ts
@@ -27,7 +27,7 @@ export class TBIncome {
   @Column('text', { name: 'cover', nullable: true, comment: '配图' })
   cover: string;
 
-  @Column('int', { name: 'price', comment: '支付金额' })
+  @Column('float', { name: 'price', comment: '支付金额' })
   price: number;
 
   @ManyToOne(() => User)
